move api_request and related methods from Handler to Utils
